+++ environment.yml
@@ -9,6 +9,7 @@ dependencies:
   - matplotlib=3.1.3
   - pandas=1.0.3
   - data_pipeline_api=0.1.0
+  - pyyaml=5.3.1
 
   # test dependencies
   - coverage=5.0
