+++ tests/unit/test_network_of_populations/mpl_images/test_visualisation__test_plotStates_single_row.png
